+++ lib/suspenders/app_builder.rb
@@ -254,10 +254,6 @@ config.public_file_server.headers = {
       copy_file "spec_helper.rb", "spec/spec_helper.rb"
     end
 
-    def configure_ci
-      template "circle.yml.erb", "circle.yml"
-    end
-
     def configure_i18n_for_test_environment
       copy_file "i18n.rb", "spec/support/i18n.rb"
     end
