+++ templates/spec_helper.rb
@@ -1,11 +1,6 @@
 if ENV.fetch("COVERAGE", false)
   require "simplecov"
 
-  if ENV["CIRCLE_ARTIFACTS"]
-    dir = File.join(ENV["CIRCLE_ARTIFACTS"], "coverage")
-    SimpleCov.coverage_dir(dir)
-  end
-
   SimpleCov.start "rails"
 end
